Refactor:
- proper DB layer with lazily loaded DB infra
- move db to persistence
- pull apart db and replay parsing code
