+++ src/ai/functions/CastReplay.py
@@ -5,8 +5,8 @@ from pydantic import BaseModel, ConfigDict, Field
 from shared import signal_queue
 from src.ai.functions.base import AIFunction
 from src.events.events import CastReplayEvent
-from src.replaydb.db import replaydb
-from src.replaydb.types import Replay
+from src.persistence.replay_store import get_replay_store
+from src.replays.types import Replay
 
 
 class CastReplayArgs(BaseModel):
@@ -35,7 +35,8 @@ def _cast_replay(
         q = Replay.unix_timestamp == int(replay_id[:10])
     else:
         q = Replay.filehash == replay_id
-    replay: Replay = replaydb.db.find_one(Replay, query=q)  # type: ignore
+    replay_store = get_replay_store()
+    replay: Replay = replay_store.db.find_one(Replay, query=q)  # type: ignore
     if not replay:
         return f"Replay with ID {replay_id} not found."
