+++ src/lib/battlenet.py
@@ -8,7 +8,7 @@ from pydantic import BaseModel, HttpUrl
 
 from config import config
 from shared import REGION_MAP
-from src.replaydb.types import ToonHandle
+from src.replays.types import ToonHandle
 
 log = logging.getLogger(f"{config.name}.{__name__}")
