+++ src/lib/sc2pulse.py
@@ -14,7 +14,7 @@ from pydantic import BaseModel, computed_field
 
 from config import config
 from src.lib.sc2client import Race as GameInfoRace
-from src.replaydb.types import ToonHandle
+from src.replays.types import ToonHandle
 from src.util import convert_enum, is_barcode
 
 log = logging.getLogger(f"{config.name}.{__name__}")
