+++ src/matchhistory.py
@@ -9,7 +9,7 @@ from pydantic import BaseModel, ConfigDict, computed_field, field_validator
 from config import config
 from shared import http_client
 from src.lib.sc2pulse import SC2PulseClient, SC2PulseCommonCharacter, SC2PulseRace
-from src.replaydb.types import ToonHandle
+from src.replays.types import ToonHandle
 
 log = logging.getLogger(f"{config.name}.{__name__}")
