+++ src/replays/__init__.py
@@ -1,0 +1,3 @@
+from src.replays.reader import ReplayReader, replay_to_dict
+
+__all__ = ["ReplayReader", "replay_to_dict"]
