+++ src/replays/types.py
@@ -18,12 +18,11 @@ from typing import (
 
 import bson
 import pydantic
-from pydantic import BaseModel, Field, ValidationError, field_validator
+from pydantic import BaseModel, Field, ValidationError
 from pydantic_core import CoreSchema, core_schema
-from pymongo import ASCENDING, IndexModel
-from pyodmongo import DbModel, Id, MainBaseModel
+from pyodmongo import DbModel, MainBaseModel
 
-from config import AIBackend, config
+from config import config
 from shared import REGION_MAP
 from src.util import time2secs
 
@@ -437,307 +436,6 @@ class AIContentPart(MainBaseModel):
     text: str
 
 
-def _data(value: Any) -> dict[str, Any]:
-    if value is None:
-        return {}
-    if isinstance(value, dict):
-        return {key: _value(item) for key, item in value.items()}
-    if hasattr(value, "model_dump"):
-        return _data(value.model_dump(mode="python"))
-    if hasattr(value, "__dict__"):
-        return {key: _value(item) for key, item in vars(value).items()}
-    return {key: _value(item) for key, item in dict(value).items()}
-
-
-def _value(value: Any) -> Any:
-    if isinstance(value, dict):
-        return _data(value)
-    if isinstance(value, (list, tuple)):
-        return [_value(item) for item in value]
-    if hasattr(value, "model_dump") or hasattr(value, "__dict__"):
-        return _data(value)
-    return value
-
-
-class AIConversation(DbModel):
-    session: Session | Id | None = None
-    trigger: AIConversationTrigger
-    status: AIConversationStatus = AIConversationStatus.active
-    closed_at: datetime | None = None
-
-    developer_instructions: str | None = None
-    handler_context: str | None = None
-    prompt_template: str | None = None
-    prompt_version: str | None = None
-
-    replay_id: str | None = None
-    map_name: str | None = None
-    opponent: str | None = None
-    twitch_user: str | None = None
-    title: str | None = None
-    metadata: dict[str, Any] = Field(default_factory=dict)
-
-    item_count: int = 0
-    last_item_at: datetime | None = None
-    last_response_id: str | None = None
-    total_input_tokens: int = 0
-    total_cached_tokens: int = 0
-    total_output_tokens: int = 0
-    total_tokens: int = 0
-    total_cost: float = 0
-
-    _collection: ClassVar = "ai_conversations"
-    _indexes: ClassVar = [
-        IndexModel(
-            [("session", ASCENDING), ("trigger", ASCENDING), ("created_at", ASCENDING)]
-        )
-    ]
-
-    @field_validator("metadata", mode="before")
-    @classmethod
-    def _default_metadata(cls, value: dict[str, Any] | None) -> dict[str, Any]:
-        if value is None:
-            return {}
-        return value
-
-    def close(self) -> None:
-        self.status = AIConversationStatus.closed
-        self.closed_at = datetime.now()
-
-    def add_item(self, item: AIConversationItem) -> None:
-        self.last_item_at = item.created_at
-        self.item_count += 1
-
-    def add_response_record(self, record: AIResponseRecord) -> None:
-        self.last_response_id = record.response_id
-        self.total_input_tokens += record.input_tokens
-        self.total_cached_tokens += record.cached_tokens
-        self.total_output_tokens += record.output_tokens
-        self.total_tokens += record.total_tokens
-        self.total_cost += record.total_cost
-
-
-class AIConversationItem(DbModel):
-    conversation: AIConversation | Id
-    session: Session | Id | None = None
-    type: AIConversationItemType
-    order: int
-
-    role: AIMessageRole | None = None
-    content: list[AIContentPart] = Field(default_factory=list)
-
-    call_id: str | None = None
-    name: str | None = None
-    arguments: dict[str, Any] | None = None
-    output: str | None = None
-
-    response_id: str | None = None
-    response_model: str | None = None
-    status: str | None = None
-    raw_item: dict[str, Any] | None = None
-
-    source: str | None = None
-    included_in_context: bool = True
-    metadata: dict[str, Any] | None = Field(default_factory=dict)
-
-    _collection: ClassVar = "ai_conversation_items"
-    _indexes: ClassVar = [
-        IndexModel(
-            [("conversation", ASCENDING), ("order", ASCENDING)],
-            unique=True,
-            partialFilterExpression={"conversation": {"$exists": True}},
-        ),
-        IndexModel([("conversation", ASCENDING), ("created_at", ASCENDING)]),
-    ]
-
-
-class AIResponseRecord(DbModel):
-    conversation: AIConversation | Id
-    session: Session | Id | None = None
-    response_id: str | None = None
-    model: str | None = None
-    status: str | None = None
-    streamed: bool = False
-
-    input_tokens: int = 0
-    cached_tokens: int = 0
-    output_tokens: int = 0
-    total_tokens: int = 0
-    input_cost: float = 0
-    cached_input_cost: float = 0
-    output_cost: float = 0
-    total_cost: float = 0
-
-    raw_usage: dict[str, Any] | None = None
-    metadata: dict[str, Any] = Field(default_factory=dict)
-
-    _collection: ClassVar = "ai_responses"
-    _indexes: ClassVar = [
-        IndexModel([("conversation", ASCENDING), ("created_at", ASCENDING)]),
-        IndexModel([("response_id", ASCENDING)], unique=True, sparse=True),
-    ]
-
-    @field_validator("metadata", mode="before")
-    @classmethod
-    def _default_metadata(cls, value: dict[str, Any] | None) -> dict[str, Any]:
-        if value is None:
-            return {}
-        return value
-
-    @classmethod
-    def from_response(
-        cls,
-        conversation: AIConversation,
-        response: Any,
-        streamed: bool = False,
-    ) -> AIResponseRecord:
-        response_data = _data(response)
-        usage = _data(response_data.get("usage"))
-        input_details = _data(usage.get("input_tokens_details"))
-        response_id = response_data.get("id") or response_data.get("response_id")
-        model = response_data.get("model")
-
-        input_tokens = int(usage.get("input_tokens") or 0)
-        cached_tokens = int(input_details.get("cached_tokens") or 0)
-        output_tokens = int(usage.get("output_tokens") or 0)
-        total_tokens = int(usage.get("total_tokens") or input_tokens + output_tokens)
-
-        pricing = config.get_model_pricing(model)
-        input_cost = max(0, input_tokens - cached_tokens) * pricing.prompt
-        cached_input_cost = cached_tokens * pricing.cached_prompt
-        output_cost = output_tokens * pricing.completion
-
-        return cls(
-            conversation=conversation,
-            session=conversation.session,
-            response_id=response_id,
-            model=model,
-            status=response_data.get("status"),
-            streamed=streamed,
-            input_tokens=input_tokens,
-            cached_tokens=cached_tokens,
-            output_tokens=output_tokens,
-            total_tokens=total_tokens,
-            input_cost=input_cost,
-            cached_input_cost=cached_input_cost,
-            output_cost=output_cost,
-            total_cost=input_cost + cached_input_cost + output_cost,
-            raw_usage=usage or None,
-        )
-
-
 class Role(str, Enum):
     user = "user"
     assistant = "assistant"
-
-
-class Metadata(DbModel):
-    replay: ReplayId
-    description: str | None = None
-    tags: List[str] = []
-    replay_summary_conversation: AIConversation | Id | None = None
-    _collection: ClassVar = "meta"
-
-
-class Alias(MainBaseModel):
-    name: str
-    portraits: list[BsonBinary] = []
-    seen_on: datetime | None = None
-
-    def __str__(self) -> str:
-        return f"{self.name}"
-
-    def __repr__(self) -> str:
-        return self.__str__()
-
-    def __eq__(self, other: object) -> bool:
-        if isinstance(other, Alias):
-            return self.name == other.name
-        elif isinstance(other, PlayerInfo):
-            if other.portrait is None:
-                return self.name == other.name
-            else:
-                return (
-                    self.name == other.name
-                    and to_bson_binary(other.portrait) in self.portraits
-                )
-        else:
-            return False
-
-
-AliasList = List[Alias]
-AliasList.__contains__ = lambda self, x: any(x == alias for alias in self)
-AliasList.__getitem__ = lambda self, x: next(alias for alias in self if alias == x)
-
-
-class PlayerInfo(DbModel):
-    id: ToonHandle = Field(...)
-    name: str
-    aliases: AliasList = []
-    toon_handle: ToonHandle
-    portrait: BsonBinary | None = None
-    portrait_constructed: BsonBinary | None = None
-    tags: List[str] | None = None
-    _collection: ClassVar = "players"
-
-    def update_aliases(self, seen_on: Optional[datetime] = None):
-        seen_on = seen_on or datetime.now()
-        if self in self.aliases:
-            return
-        for alias in self.aliases:
-            if alias.name == self.name:
-                if self.portrait and self.portrait not in alias.portraits:
-                    alias.portraits.append(self.portrait)
-                    alias.seen_on = seen_on
-                return
-
-        portraits = [self.portrait] if self.portrait else []
-
-        self.aliases.append(
-            Alias(
-                name=self.name,
-                seen_on=seen_on,
-                portraits=portraits,
-            )
-        )
-
-    def __str__(self) -> str:
-        exclude = {
-            "portrait": 1,
-            "portrait_constructed": 1,
-            "aliases.portraits": 1,
-        }
-
-        exclude_keys = convert_projection(exclude, model=PlayerInfo)
-
-        return self.model_dump_json(
-            exclude_unset=True,
-            exclude=exclude_keys,
-        )
-
-    def __repr__(self) -> str:
-        return self.__str__()
-
-
-class Session(DbModel):
-    conversations: list[AIConversation | Id] = []
-    current_conversation: AIConversation | Id | None = None
-    twitch_conversation: AIConversation | Id | None = None
-    ai_backend: AIBackend
-    session_date: datetime
-    completion_pricing: float
-    prompt_pricing: float
-    cached_prompt_pricing: float = 0
-    total_input_tokens: int = 0
-    total_cached_tokens: int = 0
-    total_output_tokens: int = 0
-    total_tokens: int = 0
-    total_cost: float = 0
-    _collection: ClassVar = "sessions"
-
-    def add_response_record(self, record: AIResponseRecord) -> None:
-        self.total_input_tokens += record.input_tokens
-        self.total_cached_tokens += record.cached_tokens
-        self.total_output_tokens += record.output_tokens
-        self.total_tokens += record.total_tokens
-        self.total_cost += record.total_cost
