+++ src/session.py
@@ -11,7 +11,6 @@ from config import AudioMode, config
 from log import log
 from src.ai import AICoach
 from src.ai.prompt import Templates
-from src.ai.state import conversation_store
 from src.contracts import MicrophoneService, TranscriberService, TTSService
 from src.events import (
     CastReplayEvent,
@@ -27,10 +26,15 @@ from src.lib.sc2client import SC2Client
 from src.lib.sc2pulse import SC2PulseClient, get_division_for_mmr
 from src.mapstats import update_map_stats
 from src.matchhistory import get_sc2pulse_match_history
+from src.persistence.conversation_store import get_conversation_store
+from src.persistence.replay_store import Metadata, get_replay_store
+from src.persistence.session_store import Session
 from src.playerinfo import resolve_replays_from_current_opponent
-from src.replaydb.db import replaydb
-from src.replaydb.types import AIConversationTrigger, Metadata, Replay, Role, Session
+from src.replays.types import AIConversationTrigger, Replay, Role
 from src.util import secs2time
+
+conversation_store = get_conversation_store()
+replay_store = get_replay_store()
 
 
 class DummyMicrophoneService(MicrophoneService):
@@ -105,7 +109,7 @@ class AISession:
             cached_prompt_pricing=config.gpt_cached_prompt_pricing,
             ai_backend=config.aibackend,
         )
-        replaydb.db.save(self.session)
+        replay_store.db.save(self.session)
 
         self.handlers = {
             TwitchChatEvent: self.handle_twitch_chat,
@@ -143,7 +147,7 @@ class AISession:
         self.session.current_conversation = value
         if value is not None and value not in self.session.conversations:
             self.session.conversations.append(value)
-        replaydb.db.save(self.session)
+        replay_store.db.save(self.session)
 
     def set_season(self):
         sc2pulse = SC2PulseClient()
@@ -159,7 +163,7 @@ class AISession:
     def update_last_replay(self, replay: Replay | None = None):
         if replay is None:
             try:
-                replay = replaydb.get_most_recent()
+                replay = replay_store.get_most_recent_for_player(config.student.name)
             except:  # noqa: E722
                 log.error("Error getting most recent replay, is DB running?")
                 sys.exit(1)
@@ -257,7 +261,7 @@ class AISession:
         and save it to the session in the DB."""
         if self.session.id is None:
             return
-        reloaded = replaydb.db.find_one(
+        reloaded = replay_store.db.find_one(
             Model=Session,
             query=Session.id == self.session.id,  # type: ignore[arg-type]
         )
@@ -267,7 +271,7 @@ class AISession:
         log.info(
             f"Total tokens: {self.session.total_tokens} (~${self.session.total_cost:.2f})"
         )
-        replaydb.db.save(self.session)
+        replay_store.db.save(self.session)
 
     def is_goodbye(self, response: str):
         return levenshtein(response[-20:].lower().strip(), "good luck, have fun") < 8
@@ -514,7 +518,7 @@ class AISession:
                 session=self.session,
             )
             self.session.twitch_conversation = self.twitch_conversation_id
-            replaydb.db.save(self.session)
+            replay_store.db.save(self.session)
         else:
             self.coach.set_active_conversation(self.twitch_conversation_id)
 
@@ -627,7 +631,7 @@ class AISession:
             schema=ReplaySummary,
         )
 
-        meta = replaydb.db.find_one(
+        meta = replay_store.db.find_one(
             Model=Metadata,
             query=Metadata.replay == replay.id,  # type: ignore[arg-type]
         )
@@ -637,4 +641,4 @@ class AISession:
         meta.description = replay_summary.description
         meta.tags = replay_summary.tags
         meta.replay_summary_conversation = self.conversation_id
-        replaydb.upsert(meta)
+        replay_store.upsert(meta)
