+++ tests/integration/test_db.py
@@ -1,15 +1,20 @@
 from datetime import datetime
 
 import pytest
+from pyodmongo.queries import eq
 
 from config import config
-from src.replaydb.db import eq, replaydb
-from src.replaydb.reader import ReplayReader
-from src.replaydb.types import AIConversation, AIConversationTrigger, Metadata, Replay
+from src.persistence.conversation_store import AIConversation
+from src.persistence.replay_store import Metadata, get_replay_store
+from src.persistence.session_store import Session
+from src.replays.reader import ReplayReader
+from src.replays.types import AIConversationTrigger, Replay
+
+replay_store = get_replay_store()
 
 
 def test_db_ready():
-    count = replaydb.replays.estimated_document_count(maxTimeMS=5000)
+    count = replay_store.replays.estimated_document_count(maxTimeMS=5000)
     assert count > 0
 
 
@@ -30,14 +35,14 @@ def test_add_metadata(replay_file):
     )
     meta.description = "This is a test description 3"
     conversation = AIConversation(trigger=AIConversationTrigger.replay_summary)
-    replaydb.db.save(conversation)
+    replay_store.db.save(conversation)
     meta.replay_summary_conversation = conversation
 
     meta.tags = ["test", "zvz", "muta"]
 
-    db_response = replaydb.upsert(meta)
+    db_response = replay_store.upsert(meta)
 
-    replaydb.db.delete(AIConversation, query=eq(AIConversation.id, conversation.id))
+    replay_store.db.delete(AIConversation, query=eq(AIConversation.id, conversation.id))
 
     assert db_response.acknowledged
     assert db_response.matched_count in {0, 1}
@@ -55,7 +60,7 @@ def test_upsert_existing_replay(replay_file):
 
     replay = reader.load_replay(replay_file)
 
-    result = replaydb.upsert(replay)
+    result = replay_store.upsert(replay)
     assert result.acknowledged
     assert result.matched_count == 1
     assert result.modified_count == 1
@@ -93,17 +98,17 @@ def test_upsert_new_replay():
         unix_timestamp=1000,
     )
 
-    result = replaydb.upsert(replay)
+    result = replay_store.upsert(replay)
     assert result.acknowledged
     assert any(new_id == v for k, v in result.upserted_ids.items())
 
-    del_result = replaydb.db.delete(Replay, query=eq(Replay.id, new_id))
+    del_result = replay_store.db.delete(Replay, query=eq(Replay.id, new_id))
 
     assert del_result.deleted_count == 1
     assert del_result.acknowledged
 
 
 def test_get_most_recent():
-    replay: Replay = replaydb.get_most_recent()
+    replay: Replay = replay_store.get_most_recent_for_player(config.student.name)
     assert any(config.student.name in player.name for player in replay.players)
     assert replay is not None
