+++ tests/llm/test_replay_timestamps.py
@@ -11,8 +11,8 @@ if not os.getenv("RUN_LIVE_OPENAI_TESTS"):
     )
 
 from coach import AISession
-from src.replaydb.reader import ReplayReader
-from src.replaydb.types import Replay
+from src.replays.reader import ReplayReader
+from src.replays.types import Replay
 
 
 @pytest.mark.parametrize(
