+++ tests/unit/test_aifunctions.py
@@ -5,7 +5,7 @@ from src.ai.functions import AddMetadata, AIFunctions
 from src.ai.functions.CastReplay import CastReplay
 from src.ai.utils import force_valid_json_string, get_clean_tags
 from src.events.events import CastReplayEvent
-from src.replaydb.types import Replay
+from src.replays.types import Replay
 
 
 @pytest.mark.parametrize(
@@ -68,14 +68,18 @@ def test_clean_tag(tags, expected):
 
 def test_castreplay_found(mocker: MockerFixture):
     mock_replay = mocker.create_autospec(Replay, instance=True)
-    mock_replaydb = mocker.patch("src.ai.functions.CastReplay.replaydb")
+    mock_replay_store = mocker.Mock()
+    mocker.patch(
+        "src.ai.functions.CastReplay.get_replay_store",
+        return_value=mock_replay_store,
+    )
     mock_signal_queue = mocker.patch("src.ai.functions.CastReplay.signal_queue")
-    mock_replaydb.db.find_one.return_value = mock_replay
+    mock_replay_store.db.find_one.return_value = mock_replay
 
     replay_id = "testhash123"
     response = CastReplay(replay_id)
 
-    mock_replaydb.db.find_one.assert_called()
+    mock_replay_store.db.find_one.assert_called()
     args, kwargs = mock_signal_queue.put.call_args
     assert isinstance(args[0], CastReplayEvent)
     assert args[0].replay == mock_replay
@@ -84,9 +88,13 @@ def test_castreplay_found(mocker: MockerFixture):
 
 
 def test_castreplay_not_found(mocker: MockerFixture):
-    mock_replaydb = mocker.patch("src.ai.functions.CastReplay.replaydb")
+    mock_replay_store = mocker.Mock()
+    mocker.patch(
+        "src.ai.functions.CastReplay.get_replay_store",
+        return_value=mock_replay_store,
+    )
     mock_signal_queue = mocker.patch("src.ai.functions.CastReplay.signal_queue")
-    mock_replaydb.db.find_one.return_value = None
+    mock_replay_store.db.find_one.return_value = None
     replay_id = "notfoundhash"
     response = CastReplay(replay_id)
     assert response == f"Replay with ID {replay_id} not found."
@@ -95,12 +103,16 @@ def test_castreplay_not_found(mocker: MockerFixture):
 
 def test_castreplay_numeric_id(mocker: MockerFixture):
     mock_replay = mocker.create_autospec(Replay, instance=True)
-    mock_replaydb = mocker.patch("src.ai.functions.CastReplay.replaydb")
+    mock_replay_store = mocker.Mock()
+    mocker.patch(
+        "src.ai.functions.CastReplay.get_replay_store",
+        return_value=mock_replay_store,
+    )
     mock_signal_queue = mocker.patch("src.ai.functions.CastReplay.signal_queue")
-    mock_replaydb.db.find_one.return_value = mock_replay
+    mock_replay_store.db.find_one.return_value = mock_replay
     numeric_id = "1746895208"
     response = CastReplay(numeric_id)
-    mock_replaydb.db.find_one.assert_called()
+    mock_replay_store.db.find_one.assert_called()
     assert numeric_id in response
     assert "Casting for" in response
     mock_signal_queue.put.assert_called()
